requirements_dev.txt
flake8
pep8
autopep8
pylint==1.7.4
pylint-flask
jsonschema
black